Improvements to function, name and arguments
This commit is a first pass at adding a number of improvements to
functions and how they are detected.

Function declarations now have a higher granularity to expose the
function keyword, the name, the parenthesis, argument definitions, and
commas within the argument definitions.

I also added a new block type called jsFuncBlock which is unique from
jsBlock that defines a block region for function definitions.

This also ultimately differentiates function definitions from
jsFuncCall.