Added jsFunctionVar and fixed up jsObjectKey
Adding specificity for variable declarations of functions to highlight the
variable differently. Only known issue is if you have a statement like
`var someVar = anotherVar = function(){};` then only anotherVar gets the
the syntax match.

I also noticed we weren't closing the zero-width match for
`jsObjectKey`, so that should be fixed now.

I also changed jsObjectKey to be `contained`, so it won't mistakenly
apply to stuff outside an object declaration.

On a general note, I am a bit concerned the computation for `jsFunctionVar`
could be an expensive regex due to it NOT being contained AND having a
lookbehind. Please report back if you notice any performance
implications of this.