Adding additional `contained`s
The intention of this commit is to perhaps improve performance, but also
clean up a bit of the how a couple regexes are used.

By adding `contained` to `jsObjectKey`, I think we might get a nice
performance boost, especially on larger files since it won't attemp that
lookbehind everywhere.

Adding `contained` to the newer regex for string templating stuff
in ES6 seems more 'accurate' or proper.