Add dot-chaining syntax to main indent logic
Most style guides depict chained methods to have a leading dot ans be
indented on level deeper then the initial statement.

This patch adds a period as a key symbol for a nested indent. In the
same fashion as the leading comma does.

It allows an indent of the following:

      var x = {
        x: 'foo',
        y: 'bar',
        z: (function() {
          return [1, 2, 3]
            .map(toFruits)
            .compact()
            .sortBy('color');
        })()
      };

This should close issue #139.