Respect 'very-magic' in 'jsAssignmentExpr' regexp
Summary:
This regex was modified in 325c0487f90707ef3e3669c1125779c7e5734037,
where the tokens `\(>\)\@!` were added.
This was supposed to be a negative lookahead for a literal `>`.
However, this didn't work because the regexp has the `\v` flag,
so the regexp actually ended up matching nothing.

Closes #333.

Test Plan:
Install the [gerw/vim-HiLinkTrace][HLT] plugin.
Write `let x = 3` in a new JavaScript file.
Place the cursor on the `x` and enter `:HLT<CR>`.
Before the patch is applied,
the `SynStack` and `HltTrace` fields of the message should be blank.
Then, apply the patch, relaunch vim, and repeat the `HLT` command;
the message should now contain `SynStack: jsAssignmentExpr->...`.

  [HLT]: https://github.com/gerw/vim-HiLinkTrace