Fix Comment Regions
Our comment detection was a bit buggy because we did not use the extend
keyword.  This meant that if the comment was contained within a region
or a match of sorts, it would get pre-maturely ended by a containing
group.  Using the extend keyword means that it will override any
containing parent matcher and still display a comment, which is how JS
actually interprets the code.