Add individual syntax groups for various parenthesis
Many of these are noise, but I don't consider function parenthesis to be so.
Rather than creating a non-function parens group, I just created a group for
each type and then linked them all to `jsParens` so that everything will
behave as it currently does for users who are using `jsParens`.