First pass at `jsObjectProp`
I had an interesting idea to disable various matches like `class` and
`require` from appearing when using `.` notation on objects.

The proposed fix is to create a special `jsObjectProp` match, with a
super high priority, that is used as a `nextgroup` for `.`

It is not assigned any colors or match groups, it simply exists to
disable other items from matching.

This is a very quick pass and I have not really tested it yet, but I
think it could be a very viable, working solution