Correctly highlight function bodies with an unnamed, generic function.
Given the following JavaScript + Flow code:

    const foo = function<T>(x: T): void {
      // arbitrary, unrelated body
      while (true) {
        console.log(x);
      }
    }

The body won't be highlighted correctly; on my machine, it's mostly in red.

This seems to be caused by incorrectly parsing the `<T>` when the
function is unnamed. Not sure why it ends up highlighting everything in
one colour.

This commit fixes the problem. I might have missed somewhere; I looked
for `jsFuncArgs` and added it to the same places.