Misc flow fixes
The biggest feature upgrade here is actual support for generics in
function calls.

Additionally I took a chance to improve other areas of generic support,
as well as fix a bunch of other random bugs I found that are too much to
document here.

It's possible these changes may have other unintended consequences, so
I will sit on this PR for a few days before merging.