eslint compiler improvements
Use `npx` instead of assuming eslint is installed globally.
(`npx` ships with `npm` since 5.2.0)

Exclude `%` from the `makeprg` command so that whole directories (or
the whole project) can be linted with `:make .` (the current file can
still be linted with `:make %`).

Use unix format since it's already included in the default
`errorformat`.

Automatically set compiler to eslint when editing a javascript file.